+++ recipes/intx/all/conanfile.py
@@ -17,13 +17,11 @@ class IntxConan(ConanFile):
     topics = ("evm", "biginteger", "arbitrary-precision", "header-only")
     package_type = "header-library"
     settings = "os", "arch", "compiler", "build_type"
+    no_copy_source = True
 
     @property
     def _min_cppstd(self):
         return 20
-
-    def layout(self):
-        basic_layout(self, src_folder="src")
 
     @property
     def _compilers_minimum_version(self):
@@ -35,13 +33,15 @@ class IntxConan(ConanFile):
             "apple-clang": "14.1",
         }
 
+    def layout(self):
+        basic_layout(self, src_folder="src")
+
     def package_id(self):
         self.info.clear()
 
     def validate(self):
-        if self.settings.get_safe("compiler.cppstd"):
+        if self.settings.compiler.get_safe("cppstd"):
             check_min_cppstd(self, self._min_cppstd)
-
         minimum_version = self._compilers_minimum_version.get(str(self.settings.compiler), False)
         if minimum_version and Version(self.settings.compiler.version) < minimum_version:
             raise ConanInvalidConfiguration(
@@ -63,3 +63,6 @@ class IntxConan(ConanFile):
     def package_info(self):
         self.cpp_info.bindirs = []
         self.cpp_info.libdirs = []
+
+        self.cpp_info.set_property("cmake_file_name", "intx")
+        self.cpp_info.set_property("cmake_target_name", "intx::intx")