+++ recipes/meson/all/conanfile.py
@@ -11,12 +11,12 @@ required_conan_version = ">=1.52.0"
 
 class MesonConan(ConanFile):
     name = "meson"
-    package_type = "application"
-    description = "Meson is a project to create the best possible next-generation build system"
-    topics = ("meson", "mesonbuild", "build-system")
+    description = "a project to create the best possible next-generation build system"
+    license = "Apache-2.0"
     url = "https://github.com/conan-io/conan-center-index"
     homepage = "https://github.com/mesonbuild/meson"
-    license = "Apache-2.0"
+    topics = ("mesonbuild", "build-system")
+    package_type = "application"
     no_copy_source = True
 
     def layout(self):
@@ -30,11 +30,7 @@ class MesonConan(ConanFile):
         self.info.clear()
 
     def source(self):
-        get(self, **self.conan_data["sources"][self.version],
-            destination=self.source_folder, strip_root=True)
-
-    def build(self):
-        pass
+        get(self, **self.conan_data["sources"][self.version], strip_root=True)
 
     def package(self):
         copy(self, "COPYING", src=self.source_folder, dst=os.path.join(self.package_folder, "licenses"))