+++ docs/changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### 10-Jul-2024 - 13:22 CEST
+
+- [feature] Add support for Conan 2.5.0 in the CI
+- [fix] Invalid configuration from tool requirement in Conan 1.x
+
 ### 22-May-2024 - 12:04 CEST
 
 - [feature] Add support for Conan 2.3.1 in the CI